change dialog state code from main vm to calendar vm
