+++ app/src/main/kotlin/com/farmer/olive/ui/MainViewModel.kt
@@ -1,11 +1,8 @@
 package com.farmer.olive.ui
 
-import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import com.farmer.home.data.CashBookRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
@@ -16,16 +13,4 @@ class MainViewModel @Inject constructor(
         OliveScreens.CashBook,
         OliveScreens.Statistics
     )
-
-    val showDialog = mutableStateOf(false)
-
-    fun sendInputCashData(
-        time: String,
-        name: String,
-        amount: String
-    ) {
-        viewModelScope.launch {
-            cashBookRepository.addCashData(time, name, amount.toIntOrNull() ?: -1)
-        }
-    }
 }
